% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcpizza-package.R
\docType{package}
\name{lcpizza-package}
\alias{lcpizza}
\alias{lcpizza-package}
\title{lcpizza: Little Caesars Pizza Data}
\description{
Data related to Little Caesars Pizza.
}
\author{
\strong{Maintainer}: Adam Shen \email{adamshen1@gmail.com}

}
\keyword{internal}
