% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcpizza-package.R
\docType{package}
\name{lcpizza-package}
\alias{lcpizza}
\alias{lcpizza-package}
\title{lcpizza: Little Caesars Pizza Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Nutrition and local pricing data for Little Caesars pizzas in Canada.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://adamoshen.github.io/lcpizza}
  \item \url{https://github.com/adamoshen/lcpizza}
  \item Report bugs at \url{https://github.com/adamoshen/lcpizza/issues}
}

}
\author{
\strong{Maintainer}: Adam Shen \email{adamshen1@gmail.com}

}
\keyword{internal}
